midx: batch object sort by first byte
The multi-pack index requires a sorted list of objects to
create a binary-searchable index of objects across multiple
pack-files. If an object appears in multiple packs, then we
select a signle copy based on the most-recent mtime among
packs containing that object.

In the case of many duplicate objects, or simply many objects,
we can speed up the de-duplication by processing objects in
batches. Using the first byte of the object ID is a natural
way to batch because the MIDX and IDX files have a fanout table
based on the first byte. This gives us a way to navigate directly
to the objects from each batch from each source.

To process a batch, create an array of MIDX entries for each object
matching that first byte value, then sort by OID (breaking ties by
recently-modified packs first). Then copy the first instance of an
object to the final object list.

Since the pack-by-pack loading happens in builtin/midx.c, move this
de-duplication to that file and add an expectation to write_midx_file()
that the input object list is sorted and de-duplicated.

Note: this commit includes fixes for the bugs introduced by a
previous version ("midx: batch object sort by first byte").

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>