worktree: add test for "HEAD" v "head" bug
On a case-insensitive operating system, "HEAD" and "head" should
resolve to the same ref. In a worktree, this does not work. A
'git rev-parse head' resolves to whatever is in ".git/HEAD" while
'git rev-parse HEAD' resolves to the correct "HEAD" file in the
appropriate worktree.

Create a test that demonstrates this bug in case-insensitie file
systems. For case sensitive file systems, create a test that
ensures we maintain that "head" is not a valid ref name.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>