remote: use generation number in ref_newer
The ref_newer() method is used by commands like 'git push' to check if a
force-push is required. It simply checks reachability between two
commits. It uses a commit walk that is slightly obfuscated by the calls
to pop_most_recent_commit() and commit_list_insert(), but essentially
all commits are walked with priority by maximum commit date until
finding the target commit or walking the entire reachable set.

In the case of an actual force-push, we walk the entire reachable set.
We can avoid this extra cost by not walking beyond the generation number
of the target commit.

If the commit-graph feature is not on, then the generation is always
GENERATION_NUMBER_INFINITY for every commit and the behavior is the
same.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>