rebase--interactive: rewrite complete_action() in C
This rewrites complete_action() from shell to C.

A new mode is added to rebase--helper (`--complete-action`), as well as
a new flag (`--autosquash`).

Finally, complete_action() is stripped from git-rebase--interactive.sh.

The original complete_action() checked twice that the todo-list had at
least one operations (the first check took place after the help message
was inserted, the second after the user edited the todo-list).  If there
are no actions, complete_action() would return with the code 2.  This is
a special case for rebase -i and -p; git-rebase.sh would then apply the
autostash, delete the state directory, and die with the message "nothing
to do".  This rewrite remove the first check (because a noop is added to
the todo-list if it is empty).  For the second case, the cleanup is
rewritten in C instead of returning 2.

As rebase -i no longer returns 2, the comment describing this behaviour
in git-rebase.sh is updated to reflect this change.

Signed-off-by: Alban Gruin <alban.gruin@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>