commit-graph: generation v3 (date ceiling)
This generation number option takes the simple approach of faking
"commit dates" being in the right order by adding an offset to the
commit date (stored in the generation number column) so the
modified commit date is always at least the maximum modified commit
date of all ancestors. This is locally computable. The one problem
with this approach is that we have a "maximum clock skew" based on
the maximum generation number we can store.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>