multi-pack-index: add 'repack' verb to builtin
The multi-pack-index helps object lookups in an environment with
many pack-files. It is better to have a single pack-file unless
constantly repacking into a large pack-file is not feasible. For
instance, if the repo is larger than half the available disk
space, then it is difficult to repack.

In an environment with many small pack-files, it is possible to
repack some of those small pack-files into slightly larger pack-
files and to expire the old ones.

Create a new 'repack' verb to the multi-pack-index builtin to
assist with this incremental repack. Later changes will offer
details on how this repack works.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>