midx: unlink unreferenced pack-files on repack
When calling 'git multi-pack-index repack', the simplest thing
we can do is delete the pack-files that are not used at all because
all of their objects appear in a newer pack-file. These are easy
to detect by counting the number of references to each pack-int-id.

For each of these pack-files, delete them and their pack-index.
Then rewrite the multi-pack-index file to not refer to those files
anymore.

Git processes that are running at the same time as this operation
will not try to read these pack-files.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>