multi-pack-index: repack small packs into big pack
Add a new 'repack' verb to the multi-pack-index builtin. The verb
inspects the multi-pack-index to see the packed size of objects
referenced by the multi-pack-index in each pack-file.

A list of objects is collected to create a new pack-file. These
objects have packed sizes that sum to at least the given --size
parameter, and each pack-file has referenced size at most that
--size. Since the multi-pack-index selects the more-recent pack
containing a duplicate object, these pack-files may be significantly
larger than their referenced size.

Inspect the packs (in alphanumeric order) until finding enough
objects from "small" packs to reach the given --size. Send the
referenced objects to a 'pack-objects' command to generate a new
pack-file and pack-index. Afterwards, generate a new
multi-pack-index. This will cause all of the pack-files we
selected to be unreferenced. They can be deleted in a subsequent
'git multi-pack-index expire' command.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>