blame: read diff config
When users use custom diff settings in their Git config, they
expect all diff algorithms to use those values. In particular,
some diffs change greatly if the "patience" algorithm is used
instead of the default.

In cases where the diff config changes the diff output, the
output of 'git blame' can also change. This alters the user's
expectations, as they look at the diff for that commit and
cannot understand how the line was changed by that commit.

Read the diff config options in 'git blame' so the diff algorithm
uses these options.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>