multi-pack-index: revise test across alternates
The test for handling multi-pack-indexes across alternates is a
bit rough. It uses a bare repo, and mixes up when the command is
run in the root and in the bare repo.

Instead, create a new, full repo "alt" and make the original repo
an alternate. Then, test that we see the multi-pack-index in the
new repo and can create a new multi-pack-index based on a new
pack-file.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>