midx: use normalized filenames
As we prepare to make the multi-pack-index files incremental,
follow the new example from the commit-graph file to normalize
the filenames to avoid issues with malformed --object-dir
strings.

Since the same normalization logic will now appear in three
or more locations, centralize the "xstrfmt then normalize"
logic into a central location.

The current patch is not based on commits that contain the
commit-graph normalization, so adapting those calls to this
new method will need to be done later.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>