midx: normalize object_dir member
The multi-pack-index stores its object directory in a local member.
This object directory may be loaded from user input, so it should
be normalized before we store it.

This has some fallout in our test suite, as the path "." is
normalized to an empty string, but then `test-tool read-midx`
cannot take an empty string for its object-dir argument. Switch
to using $(pwd) for the full directory.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>