t3311: use test_commit_bulk
One of the tests in t3311 creates 300 commits by running "test_commit"
in a loop. This requires 900 processes. Instead, we can use
test_commit_bulk to do it with only four. This improves the runtime of
the script from:

  Benchmark #1: ./t3311-notes-merge-fanout.sh --root=/var/ram/git-tests
    Time (mean Â± Ïƒ):      5.821 s Â±  0.691 s    [User: 3.146 s, System: 2.782 s]
    Range (min â€¦ max):    4.783 s â€¦  6.841 s    10 runs

to:

  Benchmark #1: ./t3311-notes-merge-fanout.sh --root=/var/ram/git-tests
    Time (mean Â± Ïƒ):      1.743 s Â±  0.116 s    [User: 1.144 s, System: 0.691 s]
    Range (min â€¦ max):    1.629 s â€¦  1.994 s    10 runs

for an average speedup of over 70%.

Unfortunately we still have to run 300 instances of "git notes add",
since the point is to test the fanout that comes from adding notes one
by one.

Signed-off-by: Jeff King <peff@peff.net>