Revert "switch: no worktree status unless real branch switch happens"
This reverts commit 65f099b3988198f0fdf3ef7a21dc01c556d21fff, which
removed logic for avoiding extra cost in "git checkout -b" in favor
of the new "git switch -c". The new switch builtin is still an
experimental feature, so stating "large repos should use 'git switch'"
is too aggressive.

When a user runs "git checkout -b <branchname>" they are pretty
clear in stating that they only want to create a branch and not
update the working directory. The difficulty comes in detecting
that we are actually in that case. That is why the
skip_merge_working_tree() method is so detailed: there are many
settings to check to ensure the arguments match as expected.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>