read-cache: don't verify path names outside sparse cone
When updating the working directory, we want to ensure we do not try
to create an invalid path according to the current filesystem. Since
8dcf39c46 (Prevent bogus paths from being added to the index,
2006-05-18), we prevent these invalid paths from being in the index
at all. However, if our repo contains one of these invalid paths,
we may want to use the sparse-checkout patterns to avoid writing
the paths but allow them in our index.

This existed in the wild already, but after 4dc42c6c1 (mingw: refuse
paths containing reserved names, 2019-12-21), some new forbidden paths
were recognized by verify_path() and these sparse-checkouts began to
fail. This also causes problems such as checking out the Linux kernel
repo on Windows: the existence of an "aux.c" file prevents _any_ of
the tree from being present on-disk, since the index cannot be created
at all.

While 8dcf39c46 was introduced to avoid invalid paths, and 4dc42c6c1
strengthened those restrictions, it makes it more difficult to
collaborate across platforms when these invalid paths are legitimate
in one filesystem and not in another.

To fix these concerns, short-circuit the call to verify_path() with
a check for the skip-worktree bit. This allows the index to contain
these invalid paths, but only if these paths are not going to be
written to the filesystem.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>