maintenance: jobs.post-command.enabled config option
The "git gc --auto" command checks the gc.auto config option to
determine if it should end quickly. However, we still have the cost
of running a process at the end of some Git commands that are
otherwise very fast. This is more of an issue on the Windows
platform due to the extra costs associated with process startup.

Create a config option, jobs.post-command.enabled, which can control
if we should run _any_ process at the end of these commands. This
defaults to 'true' to match the current behavior, but by setting the
option to 'false' we can avoid the extra call.

To test that this process is started in each case, use the trace2
event target to check for the subprocess launch.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>