maintenance: add basics of maintenance builtin
The 'git maintenance' builtin will be the main entry point for users
intending to interact with maintenance tasks. The other entry point
is "git config" if they know the "jobs.*" config space.

The current builtin is set up to have 'git maintenance run' simply
run the post_command_maintenance() method, which is a thin wrapper
around 'git gc --auto'. It does have the logic for the
jobs.post-command.enabled config setting, so we can test the builtin
in t6502-maintenance.sh by replacing the previous use of 'git
commit'.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>