maintenance: add --object-dir option
When repos use alternates to store object data that is shared by
multiple repos, those alternates are not full Git repos. This means that
we need to run maintenance tasks in a full Git repo, but point the tasks
towards the alternate.

In the multi-pack-index and commit-graph builtins, this is done with the
--object-dir option, which is replicated here.

The prune-packed builtin needs the GIT_OBJECT_DIRECTORY environment
variable, as no command-line option is available.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>