sparse-index: ensure_full_index()
The ensure_full_index() method takes the current index and updates it to
be non-sparse. It currently takes a repository, but maybe it will need
to take a repository _and_ an index_state, in the case where we might
somehow have two indexes.

Right now, ensure_full_index() is placed in enough places that almost
all index reads expand to full immediately. We will want to instead
spread these commands out in front of any method or loop that expects
the index to be full. This will provide a way to indicate that a certain
index operation is not yet "sparse aware" and these can be updated
one-by-one.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>