blame: ensure full index
The verify_working_tree_path() method in blame.c checks to see if the
given path exists at the current HEAD. With a sparse index, this path
might not be in the sparse-checkout definition and hence would be
missing at this point.

Use ensure_full_index() to expand to full for now, until we can make
this aware of sparse directory entries.

It is not enough to update index_name_pos(), since there is also the
follow-up in the case of 'pos' being negative. We will need to
investigate if index_name_pos() can be helpful by pointing to the sparse
directory entry that would be a parent directory of the given path.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>